migrate: move hooks to core (excluded from tidy for now), relocate CLI to apps/cli, add pre-commit public-include enforcement; README polish per feedback
