+++ apps/cli/cli/cache_rebuild.c
