+++ apps/cli/cli/install_hooks.c
