+++ apps/cli/cli/list.c
