+++ apps/cli/cli/list.md
