+++ apps/cli/cli/main.c
