+++ core/include/gitmind/hooks/augment.h
